+++ CHANGES.rst
@@ -52,6 +52,9 @@ Cubeviz
 
 - Fixed validation message of moment number in moment map plugin. [#1536]
 
+- Fixed ``viewer.jdaviz_helper`` returning Specviz helper instead of Cubeviz helper after Specviz
+  helper is called via ``Cubeviz.specviz``. Now ``viewer.jdaviz_helper`` always returns the Cubeviz helper. [#1546]
+
 Imviz
 ^^^^^